Add entity status icon component
